requirements.txt
pyyaml>=6.0
requests>=2.28.0
feedparser>=6.0.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.8.0
pypdf>=6.0.0
